Get rid of predicates_defined_on
